functions and nested loops

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -6,7 +6,7 @@
  */
 int main(void)
 {
-	int f1 = 0, f2 = 1, sum;
+	unsigned long f1 = 0, f2 = 1, sum;
 
 	int k;
